+++ soxspipe/commonutils/create_dispersion_map.py
@@ -508,7 +508,7 @@ class create_dispersion_map(object):
 
         # DETERMINE WHERE TO WRITE THE FILE
         home = expanduser("~")
-        outDir = self.settings["intermediate-data-root"].replace("~", home)
+        outDir = self.settings["workspace-root-dir"].replace("~", home)
         outDir += f"product/{self.recipeName}/"
         outDir = outDir.replace("//", "/")
         # Recursively create missing directories
@@ -1025,7 +1025,7 @@ class create_dispersion_map(object):
         # plt.show()
 
         home = expanduser("~")
-        outDir = self.settings["intermediate-data-root"].replace("~", home) + "/qc/pdf/"
+        outDir = self.settings["workspace-root-dir"].replace("~", home) + "/qc/pdf/"
         outDir = outDir.replace("//", "/")
         # RECURSIVELY CREATE MISSING DIRECTORIes
         if not os.path.exists(outDir):
@@ -1220,7 +1220,7 @@ class create_dispersion_map(object):
 
         # DETERMINE WHERE TO WRITE THE FILE
         home = expanduser("~")
-        outDir = self.settings["intermediate-data-root"].replace("~", home) + f"/product/{self.recipeName}"
+        outDir = self.settings["workspace-root-dir"].replace("~", home) + f"/product/{self.recipeName}"
         outDir = outDir.replace("//", "/")
         # Recursively create missing directories
         if not os.path.exists(outDir):