+++ soxspipe/recipes/soxs_nod_mode.py
@@ -110,6 +110,8 @@ class soxs_nod_mode(_base_recipe_):
 
         kw = self.kw
 
+        error = False
+
         # BASIC VERIFICATION COMMON TO ALL RECIPES
         imageTypes, imageTech, imageCat = self._verify_input_frames_basics()
 
@@ -142,6 +144,13 @@ class soxs_nod_mode(_base_recipe_):
         if arm not in self.supplementaryInput or "DISP_MAP" not in self.supplementaryInput[arm]:
             raise TypeError(
                 "Need a **** for %(arm)s - none found with the input files" % locals())
+
+        if error:
+            sys.stdout.write("\x1b[1A\x1b[2K")
+            print("# VERIFYING INPUT FRAMES - **ERROR**\n")
+            print(self.inputFrames.summary)
+            print()
+            raise TypeError(error)
 
         self.imageType = imageTypes[0]
         self.log.debug('completed the ``verify_input_frames`` method')