+++ soxspipe/recipes/soxs_order_centres.py
@@ -119,6 +119,8 @@ class soxs_order_centres(_base_recipe_):
 
         kw = self.kw
 
+        error = False
+
         # BASIC VERIFICATION COMMON TO ALL RECIPES
         imageTypes, imageTech, imageCat = self._verify_input_frames_basics()
 
@@ -151,6 +153,13 @@ class soxs_order_centres(_base_recipe_):
         if f"DISP_TAB_{arm}" not in imageCat:
             raise TypeError(
                 "Need a first guess dispersion map for %(arm)s - none found with the input files" % locals())
+
+        if error:
+            sys.stdout.write("\x1b[1A\x1b[2K")
+            print("# VERIFYING INPUT FRAMES - **ERROR**\n")
+            print(self.inputFrames.summary)
+            print()
+            raise TypeError(error)
 
         self.imageType = imageTypes[0]
         self.log.debug('completed the ``verify_input_frames`` method')
@@ -223,7 +232,7 @@ class soxs_order_centres(_base_recipe_):
             inputFrame=orderDef_image, master_bias=master_bias, dark=dark)
 
         if self.settings["save-intermediate-products"]:
-            fileDir = self.intermediateRootPath
+            fileDir = self.workspaceRootPath
             filepath = self._write(
                 self.orderFrame, fileDir, filename=False, overwrite=True, product=False)
             print(f"\nCalibrated single pinhole frame frame saved to {filepath}\n")