+++ soxspipe/recipes/tests/test_soxs_mflat.py
@@ -143,25 +143,35 @@ class test_soxs_mflat(unittest.TestCase):
 
         sofPath = "~/xshooter-pipeline-data/unittest_data/xsh/xshooter-mflat/sof/uvb_dflats.sof"
         from soxspipe.recipes import soxs_mflat
-        this = soxs_mflat(
-            log=log,
-            settings=settings,
-            inputFrames=sofPath,
-            overwrite=True
-        )
-        this.produce_product()
+        try:
+            this = soxs_mflat(
+                log=log,
+                settings=settings,
+                inputFrames=sofPath,
+                overwrite=True
+            )
+            this.produce_product()
+            assert False
+        except Exception as e:
+            assert True
+            print(str(e))
 
     def test_soxs_mflat_uvb_qflat_function(self):
 
         sofPath = "~/xshooter-pipeline-data/unittest_data/xsh/xshooter-mflat/sof/uvb_qflats.sof"
         from soxspipe.recipes import soxs_mflat
-        this = soxs_mflat(
-            log=log,
-            settings=settings,
-            inputFrames=sofPath,
-            overwrite=True
-        )
-        this.produce_product()
+        try:
+            this = soxs_mflat(
+                log=log,
+                settings=settings,
+                inputFrames=sofPath,
+                overwrite=True
+            )
+            this.produce_product()
+            assert False
+        except Exception as e:
+            assert True
+            print(str(e))
 
     def test_soxs_mflat_uvb_combined_flat_function(self):
 
@@ -175,15 +185,16 @@ class test_soxs_mflat(unittest.TestCase):
         )
         this.produce_product()
 
-    # def test_soxs_mflat_vis_long_function(self):
-    #     sofPath = "~/xshooter-pipeline-data/unittest_data/xsh/xshooter-mflat/sof/vis_long_flats.sof"
-    #     from soxspipe.recipes import soxs_mflat
-    #     this = soxs_mflat(
-    #         log=log,
-    #         settings=settings,
-    #         inputFrames=sofPath
-    #     )
-    #     this.produce_product()
+    def test_soxs_mflat_vis_long_function(self):
+        sofPath = "~/xshooter-pipeline-data/unittest_data/xsh/xshooter-mflat/sof/vis_long_flats.sof"
+        from soxspipe.recipes import soxs_mflat
+        this = soxs_mflat(
+            log=log,
+            settings=settings,
+            inputFrames=sofPath,
+            overwrite=True
+        )
+        this.produce_product()
 
     # def test_soxs_mflat_vis_short_function(self):
     #     sofPath = "~/xshooter-pipeline-data/unittest_data/xsh/xshooter-mflat/sof/vis_short_flats.sof"